localdb module redesign
- database module is now the one that interfaces directly with gluesql

- client module aggregates the database module

- table and index creation scripts moved to a file named "create" in the "sql"
submodule
