+++ src/localdb/mod.rs
@@ -1,3 +1,3 @@
-pub mod constants;
+mod sql;
+mod database;
 pub mod client;
-pub mod database;
